Update Product punc

+++ dotNet5783_6686_9339/BL/BO/ProductForList .cs
@@ -9,7 +9,7 @@ public class ProductForList
 {
     public int ProductId { get; set; }
     public string ProductName { get; set; }
-    public int ProductPrice { get; set; }
+    public double ProductPrice { get; set; }
     public ProductCategory Category { get; set; }
     public override string ToString() =>
 $@"
